An Assignment that required the creation of a Change Breakdown program and a Password Generator
The program would essential take the total cost of all your products as well as the amount you paid as inputs, and would return the exact change you need  as well as the number and amount of currency denominations ( Notes and coins) that you would get.

(Currently only returns change in the currency *Ghana Cedi)

The password Generator automatically produces a random password based on conditional inputs that are given. these include the minimum password length, the Minimum number of Uppercase letters, the minimum number of digits and the minimum number of special characters as well as a string which would contain the special characters that are desired to be used.
(*there are some missing input checks to prevent runtime and logic errors.)

(* The method for the verification of the password conditions might not be sufficient, would update it later with a better method for verification.